ai-analytics/requirements.txt
# EcoVerse AI/ML Analytics Requirements
# Core ML and Data Science
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# API and Web Framework
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0

# Data Processing and Analysis
python-dateutil==2.8.2
pytz==2023.3

# Environment and Configuration
python-dotenv==1.0.0

# Optional: Enhanced ML capabilities
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.13.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0

# Additional dependencies for integration
schedule==1.2.0
